container-compatible update

+++ update.sh
@@ -43,62 +43,49 @@ if [ -f ./sbin/eva-tinyapi ]; then
     exit 1
   fi
 
-  if [ "$CURRENT_BUILD" -ge "$BUILD" ]; then
+  if [ ! -f /.eva_container ] && [ "$CURRENT_BUILD" -ge "$BUILD" ]; then
     echo "Your build is ${CURRENT_BUILD}, this script can update EVA ICS to ${BUILD} only"
     exit 1
   fi
 fi
 
-rm -rf _update
-
-echo "- Starting update to ${VERSION} build ${BUILD}"
-
-mkdir -p _update
-
-if ! touch _update/test; then
-  echo "Unable to write. Read-only file system?"
-  exit 1
-fi
-
-echo "- Downloading new version tarball"
-
-cd _update || exit 1
-
-if [ -f ../eva-${VERSION}-${BUILD}.tgz ]; then
-  cp ../eva-${VERSION}-${BUILD}.tgz .
-else
-  curl -L ${EVA_REPOSITORY_URL}/${VERSION}/nightly/eva-${VERSION}-${BUILD}.tgz \
-    -o eva-${VERSION}-${BUILD}.tgz || exit 1
-fi
-
-echo "- Extracting"
-
-tar xzf eva-${VERSION}-${BUILD}.tgz || exit 1
-
-cd ..
-
-echo "- Stopping everything"
-
-./sbin/eva-control stop
-
-if [ -f ./runtime/uc_cvars.json ] || [ -f ./runtime/lm_cvars.json ] || [ -f ./runtime/sfa_cvars.json ]; then
-  echo "EVA ICS obsolete configuration found. Checking..."
-  [ -f ./venv/bin/python3 ] && PYTHON=./venv/bin/python || PYTHON=./python3/bin/python3
-  $PYTHON ./_update/eva-${VERSION}/cli/convert-legacy-configs.py check --dir $(pwd) || exit 3
-  if [ -d ./venv ]; then
-    echo "Removing new-style venv in case of a failed previous update..."
-    rm -rf ./venv
+if [ ! -f /.eva_container ]; then
+  rm -rf _update
+  echo "- Starting update to ${VERSION} build ${BUILD}"
+  mkdir -p _update
+  if ! touch _update/test; then
+    echo "Unable to write. Read-only file system?"
+    exit 1
   fi
-fi
-
-echo "- Installing missing modules"
-
-./_update/eva-${VERSION}/install/build-venv . || exit 2
-
-if [ "$CHECK_ONLY" = 1 ]; then
-  echo
-  echo "Checks passed, venv updated. New version files can be explored in the _update dir"
-  exit 0
+  echo "- Downloading new version tarball"
+  cd _update || exit 1
+  if [ -f ../eva-${VERSION}-${BUILD}.tgz ]; then
+    cp ../eva-${VERSION}-${BUILD}.tgz .
+  else
+    curl -L ${EVA_REPOSITORY_URL}/${VERSION}/nightly/eva-${VERSION}-${BUILD}.tgz \
+      -o eva-${VERSION}-${BUILD}.tgz || exit 1
+  fi
+  echo "- Extracting"
+  tar xzf eva-${VERSION}-${BUILD}.tgz || exit 1
+  cd ..
+  echo "- Stopping everything"
+  ./sbin/eva-control stop
+  if [ -f ./runtime/uc_cvars.json ] || [ -f ./runtime/lm_cvars.json ] || [ -f ./runtime/sfa_cvars.json ]; then
+    echo "EVA ICS obsolete configuration found. Checking..."
+    [ -f ./venv/bin/python3 ] && PYTHON=./venv/bin/python || PYTHON=./python3/bin/python3
+    $PYTHON ./_update/eva-${VERSION}/cli/convert-legacy-configs.py check --dir $(pwd) || exit 3
+    if [ -d ./venv ]; then
+      echo "Removing new-style venv in case of a failed previous update..."
+      rm -rf ./venv
+    fi
+  fi
+  echo "- Installing missing modules"
+  ./_update/eva-${VERSION}/install/build-venv . || exit 2
+  if [ "$CHECK_ONLY" = 1 ]; then
+    echo
+    echo "Checks passed, venv updated. New version files can be explored in the _update dir"
+    exit 0
+  fi
 fi
 
 echo "- Removing obsolete files and folders"
@@ -107,56 +94,47 @@ for o in ${OBS}; do
   rm -rf "${o}"
 done
 
-if [ ! -d ./backup ]; then
-  mkdir ./backup
-  chmod 700 ./backup
-fi
-
-rm -rf ./python3
-
-echo "- Installing new files"
-
-rm -f _update/eva-${VERSION}/ui/index.html
-rm -f _update/eva-${VERSION}/update.sh
-
-[ -L ./lib/eva/features ] && rm -f ./lib/eva/features
-[ -L ./lib/eva/uc/drivers ] && rm -f ./lib/eva/uc/drivers
-[ -L ./lib/eva/lm/extensions ] && rm -f ./lib/eva/lm/extensions
-
-cp -rf _update/eva-${VERSION}/* . || exit 1
-
-(cd ./lib/eva && ln -sf ../../runtime/plugins) || exit 1
-
-rm -f bin/eva-shell
-ln -sf eva bin/eva-shell
-
-./install/mklinks || exit 1
-
-[ -x ./sbin/registry-control ] && ./sbin/registry-control stop
-
-./install/install-yedb || exit 2
-
-if [ -f /etc/systemd/system/eva-ics.service ]; then
-  if ! grep eva-ics-registry /etc/systemd/system/eva-ics.service >& /dev/null; then
-    if [ "$(id -u)" = "0" ]; then
-      echo "- Installing EVA ICS registry service"
-      PREFIX=$(pwd)
-      sed "s|/opt/eva|${PREFIX}|g" ./etc/systemd/eva-ics-registry.service > /etc/systemd/system/eva-ics-registry.service
-      sed "s|/opt/eva|${PREFIX}|g" ./etc/systemd/eva-ics.service > /etc/systemd/system/eva-ics.service
-      if systemctl -a |grep eva-ics|grep active >& /dev/null ; then
-        echo "- Enabling EVA ICS registry service"
-        systemctl enable eva-ics-registry.service
-        systemctl daemon-reload
+if [ ! -f /.eva_container ]; then
+  if [ ! -d ./backup ]; then
+    mkdir ./backup
+    chmod 700 ./backup
+  fi
+  rm -rf ./python3
+  echo "- Installing new files"
+  rm -f _update/eva-${VERSION}/ui/index.html
+  rm -f _update/eva-${VERSION}/update.sh
+  [ -L ./lib/eva/features ] && rm -f ./lib/eva/features
+  [ -L ./lib/eva/uc/drivers ] && rm -f ./lib/eva/uc/drivers
+  [ -L ./lib/eva/lm/extensions ] && rm -f ./lib/eva/lm/extensions
+  cp -rf _update/eva-${VERSION}/* . || exit 1
+  (cd ./lib/eva && ln -sf ../../runtime/plugins) || exit 1
+  rm -f bin/eva-shell
+  ln -sf eva bin/eva-shell
+  ./install/mklinks || exit 1
+  [ -x ./sbin/registry-control ] && ./sbin/registry-control stop
+  ./install/install-yedb || exit 2
+  if [ -f /etc/systemd/system/eva-ics.service ]; then
+    if ! grep eva-ics-registry /etc/systemd/system/eva-ics.service >& /dev/null; then
+      if [ "$(id -u)" = "0" ]; then
+        echo "- Installing EVA ICS registry service"
+        PREFIX=$(pwd)
+        sed "s|/opt/eva|${PREFIX}|g" ./etc/systemd/eva-ics-registry.service > /etc/systemd/system/eva-ics-registry.service
+        sed "s|/opt/eva|${PREFIX}|g" ./etc/systemd/eva-ics.service > /etc/systemd/system/eva-ics.service
+        if systemctl -a |grep eva-ics|grep active >& /dev/null ; then
+          echo "- Enabling EVA ICS registry service"
+          systemctl enable eva-ics-registry.service
+          systemctl daemon-reload
+        fi
+      else
+        echo "- WARNING! EVA ICS sevice is installed but update isn't launched under root"
+        echo "- WARNING! Please install new eva-ics.service and eva-ics-registry.service manually"
+        sleep 3
       fi
-    else
-      echo "- WARNING! EVA ICS sevice is installed but update isn't launched under root"
-      echo "- WARNING! Please install new eva-ics.service and eva-ics-registry.service manually"
-      sleep 3
     fi
   fi
+  ./sbin/registry-control start || exit 2
 fi
 
-./sbin/registry-control start || exit 2
 
 ./install/import-registry-schema || exit 8
 ./install/import-registry-defaults || exit 8
@@ -268,19 +246,20 @@ if [ "$SFA_SETUP" == "1" ]; then
   ./sbin/eva-update-tables sfa || exit 1
 fi
 
-echo "- Cleaning up"
-
-rm -rf _update
-
-CURRENT_BUILD=$(./sbin/eva-tinyapi -B)
-
-if [ "$CURRENT_BUILD" == "${BUILD}" ]; then
-  echo "- Current build: ${BUILD}"
-  echo "---------------------------------------------"
-  echo "Update completed. Starting everything back"
-  ./sbin/registry-control start
-  ./sbin/eva-control start
+if [ ! -f /.eva_container ]; then
+  echo "- Cleaning up"
+  rm -rf _update
+  CURRENT_BUILD=$(./sbin/eva-tinyapi -B)
+  if [ "$CURRENT_BUILD" == "${BUILD}" ]; then
+    echo "- Current build: ${BUILD}"
+    echo "---------------------------------------------"
+    echo "Update completed. Starting everything back"
+    ./sbin/eva-control start
+  else
+    echo "Update failed"
+    exit 1
+  fi
 else
-  echo "Update failed"
-  exit 1
+  echo "---------------------------------------------"
+  echo "Container configuration updated successfully"
 fi
